refactor(query): extract stream event mapping into runner/stream.rs
